+++ react/components/MinicartFreeshipping.tsx
@@ -1,6 +1,7 @@
 import type { FunctionComponent } from 'react'
 import React, { useEffect, useCallback, useState } from 'react'
 import { useOrderForm } from 'vtex.order-manager/OrderForm'
+import { useRuntime } from 'vtex.render-runtime'
 import { useQuery } from 'react-apollo'
 import { FormattedMessage } from 'react-intl'
 import { FormattedCurrency } from 'vtex.format-currency'
@@ -9,9 +10,16 @@ import styles from './MinicartFreeshipping.css'
 import AppSettings from './minicartbarSettings.graphql'
 
 interface SettingsProps {
-  settings: Settings
+  settings: BindingBoundedSettings
 }
+
+interface BindingBoundedSettings extends Settings {
+  bindingBounded?: boolean
+  settings?: [Settings]
+}
+
 interface Settings {
+  bindingId: string
   freeShippingAmount: number
 }
 
@@ -20,18 +28,32 @@ type ValueTypes = 'Discounts' | 'Items'
 const MinimumFreightValue: FunctionComponent<SettingsProps> = ({
   settings,
 }) => {
+  const { binding } = useRuntime()
   const [shippingFreePercentage, setShippingFreePercentage] = useState(0)
   const [differenceBetwenValues, setDifferenceBetwenValues] = useState(0)
+  const [freeShippingAmount, setFreeShippingAmount] = useState(0)
   const {
     orderForm: { totalizers },
   } = useOrderForm()
 
+  useEffect(() => {
+    if (settings.bindingBounded) {
+      const findAmountForBinding = settings.settings?.find(
+        item => item.bindingId === binding?.id
+      )?.freeShippingAmount
+
+      if (findAmountForBinding) setFreeShippingAmount(findAmountForBinding)
+    } else {
+      setFreeShippingAmount(settings.freeShippingAmount)
+    }
+  }, [binding])
+
   const handleUpdateMinicartValue = useCallback(
     val => {
-      setShippingFreePercentage(Math.round(val / settings.freeShippingAmount))
-      setDifferenceBetwenValues(settings.freeShippingAmount - val / 100)
+      setShippingFreePercentage(Math.round(val / freeShippingAmount))
+      setDifferenceBetwenValues(freeShippingAmount - val / 100)
     },
-    [settings.freeShippingAmount]
+    [freeShippingAmount]
   )
 
   const getValues = (idValue: ValueTypes): number =>
@@ -45,21 +67,23 @@ const MinimumFreightValue: FunctionComponent<SettingsProps> = ({
 
   return (
     <div className={styles.freigthScaleContainer}>
-      {differenceBetwenValues === settings.freeShippingAmount ? (
+      {differenceBetwenValues === freeShippingAmount ? (
         <div className={styles.text0}>
           <FormattedMessage id="store/minicartbar.text0" />
           <FormattedCurrency value={Math.max(0, differenceBetwenValues)} />!
         </div>
       ) : (
         <>
-          <span>
-            <div className={styles.text1}>
-              <FormattedMessage id="store/minicartbar.text1" />
-              <span className={styles.text2}>
-                <FormattedMessage id="store/minicartbar.text2" />
-              </span>
-            </div>
-          </span>
+          {differenceBetwenValues > 0 ? (
+            <span>
+              <div className={styles.text1}>
+                <FormattedMessage id="store/minicartbar.text1" />
+                <span className={styles.text2}>
+                  <FormattedMessage id="store/minicartbar.text2" />
+                </span>
+              </div>
+            </span>
+          ) : null}
           <div className={styles.sliderContainer}>
             <div
               className={styles.barContainer}
@@ -96,13 +120,24 @@ const MinimumFreightValue: FunctionComponent<SettingsProps> = ({
 
 const MinicartFreeshipping: FunctionComponent = () => {
   const { data } = useQuery(AppSettings, { ssr: false })
+  const { binding } = useRuntime()
 
   if (!data?.appSettings?.message) return null
 
   const settings = JSON.parse(data.appSettings.message)
 
-  if (!settings.freeShippingAmount) {
+  if (!settings.bindingBounded && !settings.freeShippingAmount) {
     console.warn('No Free Shipping amount set')
+
+    return null
+  }
+
+  const isAmountSetForBinding = settings.settings?.find(
+    item => item.bindingId === binding?.id
+  )?.freeShippingAmount
+
+  if (settings.bindingBounded && !isAmountSetForBinding) {
+    console.warn('No Free Shipping amounts for multi binding store set')
 
     return null
   }